Moved controller and controllerserver to jujuclient.
